test(.net): replace dotnet-test-as-helper with standalone console app
Root cause of the QueueWatcher CI failures: subprocess "workers" were
spawned via `dotnet test --filter ...`, which forces full vstest host
startup (~30s on macos-14 ARM CI). That blew past WaitReady's 30s
budget intermittently and made the apparent failure shape "worker
never started" — a symptom of the test harness, not of any real bug.

Fix:
- New tests/Honker.QueueHelper console app holds the worker/writer
  logic, builds via ProjectReference from Honker.Tests, and starts in
  sub-second time.
- Tests now spawn the helper DLL directly (`dotnet helper.dll worker`).
- QueueWatcherBackendProcessHelper xunit method is gone — its body
  moved to the helper's Program.cs, which is the only place it ever
  ran in production anyway.
- Parent tests now enqueue an explicit { stop = true } job after the
  data jobs so worker exits deterministically; the per-helper CTS is
  now a 30s safety net rather than the load-bearing exit signal.
