+++ packages/honker-dotnet/tests/Honker.QueueHelper/Honker.QueueHelper.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <IsPackable>false</IsPackable>
+    <RootNamespace>Honker.QueueHelper</RootNamespace>
+    <AssemblyName>Honker.QueueHelper</AssemblyName>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Honker\Honker.csproj" />
+  </ItemGroup>
+
+</Project>
